+++ recipe/deploy/update_code.php
@@ -15,6 +15,7 @@ task('deploy:update_code', function () {
     $branch = get('branch');
     $git = get('bin/git');
     $gitCache = get('git_cache');
+    $recursive = get('git_recursive', true) ? '--recursive' : '';
     $depth = $gitCache ? '' : '--depth 1';
     $options = [
         'tty' => get('git_tty', false),
@@ -52,14 +53,14 @@ task('deploy:update_code', function () {
 
     if ($gitCache && has('previous_release')) {
         try {
-            run("$git clone $at --recursive -q --reference {{previous_release}} --dissociate $repository  {{release_path}} 2>&1", $options);
+            run("$git clone $at $recursive -q --reference {{previous_release}} --dissociate $repository  {{release_path}} 2>&1", $options);
         } catch (RuntimeException $exc) {
             // If {{deploy_path}}/releases/{$releases[1]} has a failed git clone, is empty, shallow etc, git would throw error and give up. So we're forcing it to act without reference in this situation
-            run("$git clone $at --recursive -q $repository {{release_path}} 2>&1", $options);
+            run("$git clone $at $recursive -q $repository {{release_path}} 2>&1", $options);
         }
     } else {
         // if we're using git cache this would be identical to above code in catch - full clone. If not, it would create shallow clone.
-        run("$git clone $at $depth --recursive -q $repository {{release_path}} 2>&1", $options);
+        run("$git clone $at $depth $recursive -q $repository {{release_path}} 2>&1", $options);
     }
 
     if (!empty($revision)) {